Reorganize frontend
