+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Navbar from './components/Navbar';
-import RateTeamContent from './components/RateTeamContent';
+import Navbar from './components/navbar/Navbar';
+import RateTeamContent from './components/rateTeamPage/RateTeamContent';
 import "./App.css";
 
 function App() {
