+++ frontend/src/components/navbar/Navbar.css
@@ -1,7 +1,7 @@
 @font-face {
     font-family: "MervaleScript-Regular";
     src: local("MervaleScript-Regular"),
-     url("../fonts/MervaleScript-Regular.ttf") format("truetype");
+     url("src/fonts/MervaleScript-Regular.ttf") format("truetype");
     font-weight: bold;
 }
