+++ frontend/src/components/rateTeamPage/RateTeamContent.css
