+++ frontend/src/components/rateTeamPage/doubloonAdder/DoubloonsAdder.css
