+++ frontend/src/components/rateTeamPage/progressBar/ProgressBar.css
