+++ frontend/src/components/rateTeamPage/progressBar/ProgressBar.tsx
