+++ frontend/src/components/rateTeamPage/teamMember/TeamMemberView.css
