+++ frontend/src/components/rateTeamPage/timetable/TimetableType.ts
